Tag Stories in FeedReader (#1171)
* Tag Stories in FeedReader

With the FeedReader, it's possible to have multiple feeds. With multiple feeds, all stories are jumbled up in the same
list. One cannot know the source of a story at first glance. The way this possible is the to use the `t` keyboard
shortcut which cycles the showType through title, link, and content. It is better UX to have some tag showing the origin
of a story.

This change introduces a tag on the story title. This tag is the same as the Feed.Title. This change is backward
compatible. If a feed does not come with a title, this change yields itself to the previous implementation.

Fixes wtfutil/wtf#1138

* Format with gofmt