fix issue with pocket client
The default marshaling behavior of `json.Marshal` returns a string
`null` for a `nil` interface. This makes the DefaultHTTPClient provide
the string `null` as the body where none was expected.
Without this, the cloudflare server returns a 403 Bad Request, causing
the widget to show an empty pane.

Additionally set a custom user-agent to identify wtfutil. in-line with
the reddit submodule.