Make Feed Reader work with Reddit feeds
Adds a `disableHTTP2` setting (default: false) which disables the HTTP/2
protocol in the client used for fetching RSS feeds.

An optional `userAgent` setting is also introduced, which makes it possible
to configure the User-Agent used by the RSS client.

These two options combined fixes the problem where Reddit returns a
`403 Forbidden` response because of Reddit's aggressive anti-bot/automation
protection.

By disabling HTTP/2 and changing the User-Agent, this problem is fixed (for now).